[ui] 2 - Moved relative strength heat map prototype to features/rs/comps/dashboard component.  Move associated files to rs feature directory
