+++ src/app/features/rs/common/constants-rs.ts
