+++ src/app/features/rs/common/interfaces-rs.ts
