+++ src/app/features/rs/comps/dashboard/dashboard.component.scss
@@ -1,0 +1,59 @@
+
+
+:host {
+
+    width: 100%;
+    height: 100%;
+    margin: 50px;
+
+    .table-container {
+        display: flex;
+        justify-content: flex-start;
+        align-items: flex-start;
+
+        .symbols-column {
+            display: flex;
+            flex-direction: column;
+            justify-content: flex-start;
+            align-items: center;
+        }
+
+        .rows-container {
+            display: flex;
+            flex-direction: column;
+            justify-content: flex-start;
+            align-items: flex-start;
+            overflow-x: scroll;
+            width: 100%;
+    
+            .row {
+                display: flex;
+                justify-content: flex-end;
+                align-items: center;
+            }
+        }
+    }
+
+    .cell {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        height: 50px;
+        width: 50px;
+        border: 1px solid black;
+    }
+
+    .column-cell {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        height: 50px;
+        width: 100px;
+        border: 1px solid black;
+    }
+
+    .text {
+        font-size: 12px;
+        margin: 0;
+    }
+}
