+++ src/app/features/rs/comps/dashboard/dashboard.component.spec.ts
@@ -20,4 +20,17 @@ describe('DashboardComponent', () => {
   it('should create', () => {
     expect(component).toBeTruthy();
   });
+
+  it(`should have the 'rel-str' title`, () => {
+    const fixture = TestBed.createComponent(DashboardComponent);
+    const app = fixture.componentInstance;
+    expect(app.title).toEqual('rel-str');
+  });
+
+  it('should render title', () => {
+    const fixture = TestBed.createComponent(DashboardComponent);
+    fixture.detectChanges();
+    const compiled = fixture.nativeElement as HTMLElement;
+    // expect(compiled.querySelector('h1')?.textContent).toContain('Hello, rel-str');
+  });
 });
