+++ src/app/features/rs/data/AMAT.ts
@@ -1,4 +1,4 @@
-import { StockDatum, StockData } from "./../../app/interfaces";
+import { StockDatum, StockData } from "../common/interfaces-rs";
 
 export const AMAT: StockDatum[] = [
     {"45037": 112.9},
