+++ src/app/features/rs/data/MSFT.ts
@@ -1,4 +1,4 @@
-import { StockDatum, StockData } from "./../../app/interfaces";
+import { StockDatum, StockData } from "../common/interfaces-rs";
 
 export const MSFT: StockDatum[] = [
     {"45037": 286.14},
