+++ src/app/features/rs/data/QQQ.ts
@@ -1,4 +1,4 @@
-import { StockDatum, StockData } from "./../../app/interfaces";
+import { StockDatum, StockData } from "../common/interfaces-rs";
 
 export const QQQ: StockDatum[] = [
     {"45037": 318.84},
