+++ src/app/features/rs/data/stocks.ts
